+++ core/vil1/file_formats/vil1_viff.cxx
@@ -389,7 +389,7 @@ bool vil1_viff_generic_image::check_endian()
 
 vil1_image vil1_viff_generic_image::get_plane(unsigned int plane) const
 {
-  assert(plane < planes_);
+  assert((int)plane < planes_);
   vcl_cerr << "FIXME: this should be an adapter that shifts to the plane asked for\n";
   return const_cast<vil1_viff_generic_image*>(this);
 }
